requirements.txt
django==1.11.8
line-bot-sdk==1.5.0
django-heroku==0.2.0
requests==2.18.4
